refactor: Terms.tsx 삭제 및 컴포넌트 화
