+++ src/routes/routes.tsx
@@ -3,7 +3,6 @@ import RootLayout from "@/layout/RootLayout";
 import Home from "@/pages/Home";
 import LoginPage from "@/pages/LoginPage";
 import RegisterPage from "@/pages/RegisterPage";
-import Terms from "@/pages/Terms.tsx";
 import OnboardingPage from "@/pages/OnboardingPage";
 import Places from "@/pages/Places";
 import Profile from "@/pages/Profile";
@@ -25,9 +24,6 @@ const router = createBrowserRouter([
       },
       { path: pagePath.REGISTER,
         element: <RegisterPage />
-      },
-      { path: pagePath.TERMS,
-        element: <Terms />
       },
       {
         path: pagePath.PLACES,
